+++ src/xmpp/xmpp-im/httpfileupload.cpp
@@ -442,13 +442,13 @@ HttpFileUpload *HttpFileUploadManager::upload(const QString &srcFilename, const 
 {
     auto f = new QFile(srcFilename);
     f->open(QIODevice::ReadOnly);
-    auto hfu = upload(f, size_t(f->size()), dstFilename, mType);
+    auto hfu = upload(f, f->size(), dstFilename, mType);
     connect(hfu, &HttpFileUpload::finished, this, [f]() { f->close(); });
     f->setParent(hfu);
     return hfu;
 }
 
-HttpFileUpload *HttpFileUploadManager::upload(QIODevice *source, size_t fsize, const QString &dstFilename,
+HttpFileUpload *HttpFileUploadManager::upload(QIODevice *source, quint64 fsize, const QString &dstFilename,
                                               const QString &mType)
 {
     auto                   hfu  = new HttpFileUpload(d->client, source, fsize, dstFilename, mType);
